Client side generation of graph
